% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphaValueEstimation.R
\name{alphaValueEstimation}
\alias{alphaValueEstimation}
\title{Estimates AlphaValues based on window offsets for a Genomic Range}
\usage{
alphaValueEstimation(m, offsets)
}
\arguments{
\item{m}{Granges including Total,Methylated Counts}

\item{offsets}{list of offsets describing the datapoints to be used to calc the alpha values for each parameter}
}
\value{
estimated alpha values
}
\description{
The list offsets musst be in the format:
 list(offsetConfig1->list(chr1,2,..),offsetConfig2->list(chr1,2,..))
 Offsets are relative to the datapoint position. (-1,-2,-3,0,1,2,3)
}
