% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEPICidats.R
\name{buildMethGrangesFromRnbSet}
\alias{buildMethGrangesFromRnbSet}
\title{Builds a Methylation GRange object from RnbSet that can be used for PMD detection.}
\usage{
buildMethGrangesFromRnbSet(rnb.obj, samples = NULL)
}
\arguments{
\item{rnb.obj}{a rnbeads object that represents an EPIC array experiment}

\item{samples}{the sample that will be used to create tbe Meth. Grange. If a vector is supplied the mean Meth over all specified samples is taken.}
}
\value{
grange Object with Methylation and Total counts
}
\description{
Builds a Methylation GRange object from RnbSet that can be used for PMD detection.
}
