% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateAlphaDistr.R
\name{calculateAlphaDistr}
\alias{calculateAlphaDistr}
\title{Estimates AlphaValues based on window offsets for a Chromosome'}
\usage{
calculateAlphaDistr(M, T, offsets)
}
\arguments{
\item{offsets}{list of relative offsets describing the datapoints to be used for alpha estimation}

\item{m}{Granges of one particular chromosome. Metadata musst include Total,Methylated Counts}
}
\value{
estimated alpha values
}
\description{
Estimates AlphaValues based on window offsets for a Chromosome'
}
