% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNNoffsets.R
\name{calculateKNNs}
\alias{calculateKNNs}
\title{Calculates the KNNs for each datapoint based on the genomic distance}
\usage{
calculateKNNs(m, ks)
}
\arguments{
\item{m}{GenomicRange containg EPIC datapoints with Methylation, Total Value Counts}

\item{ks}{vector of values for K}
}
\value{
list(knnValue->list(chrs->FNN knn-objects))
consider reading FNN::get.knnx documentation
}
\description{
Calculates the KNNs for each datapoint based on the genomic distance
}
