% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNNoffsets.R
\name{calculateKNNsOffsets}
\alias{calculateKNNsOffsets}
\title{Calculates the offsets for each datapoint used to estimate it's alpha value}
\usage{
calculateKNNsOffsets(knns.list, cutoff = NULL, q = 1)
}
\arguments{
\item{knns.list}{list(knnValue->list(chr)->KNNs)}

\item{cutoff}{distance cutoff at which datapoints should be discarded}

\item{q}{quantile based distance cutoff at which datapoints should be discarded, q=1 -> disabled}
}
\value{
list(knnValue->list(chrs->vector(list(dp->offsets))))
}
\description{
Calculates the offsets for each datapoint used to estimate it's alpha value
}
