% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGRangesObjectPMDSegmentation.R
\name{createGRangesObjectPMDSegmentation}
\alias{createGRangesObjectPMDSegmentation}
\title{Generates a genome segmentation including PMDs,nonPMDs and their unsure counterparts.'}
\usage{
createGRangesObjectPMDSegmentation(m.list, y.list, seqLengths)
}
\arguments{
\item{y.list}{prediction object of library(mhsmm).}

\item{seqLengths}{Seqlenghts objects representing the assembly sequence lengths of the chromosomes.}

\item{m}{GRange representing the EPIC datapoints. Metadata musst include Total,Methylated Counts}
}
\value{
GRange genome segmentation with PMDs,nonPMDs and their unsure counterparts.
}
\description{
Generates a genome segmentation including PMDs,nonPMDs and their unsure counterparts.'
}
