% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNNoffsets.R
\name{getHighVariableDistNNs}
\alias{getHighVariableDistNNs}
\title{Filters for data points that have a high variance in distance to their NNs}
\usage{
getHighVariableDistNNs(nn)
}
\arguments{
\item{nn}{neighrest neighbor FNN::get.knnx object}
}
\value{
indicies of the detected data points
}
\description{
Filters for datapoints that are out of the .25-.75 quantile in regards to 
their variance in distance
}
