% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMDviterbiSegmentation.R
\name{PMDviterbiSegmentation}
\alias{PMDviterbiSegmentation}
\title{Predicts PMDs/nPMD segments based on HiddenMarkovModel. Selection by ML using viterbi selection.}
\usage{
PMDviterbiSegmentation(alphas, hmm.model, markLowDensitySegm = T, m.list)
}
\arguments{
\item{alphas}{estimated alpha values for each datapoint. format:list(chr->vector(alphaValues))}

\item{hmm.model}{trained Hidden Markov Model}

\item{markHighDPdistVarSegms}{mark segments with high variance in pairwise dp distance}

\item{knns}{list(k->list(chrs->FNN::get.knnx objs)), only needed if markHighDPdistVarSegms is TRUE}
}
\value{
prediction of PMD/nPMD segments for each chromosome
}
\description{
Predicts PMDs/nPMD segments based on HiddenMarkovModel. Selection by ML using viterbi selection.
}
