% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEPICidats.R
\name{readEPIC_idats}
\alias{readEPIC_idats}
\title{Reads EPIC arrays idat files with RnBeads default pipeline}
\usage{
readEPIC_idats(sample.sheet, idat.dir, sample.col.name, preprocess = T)
}
\arguments{
\item{sample.sheet}{a csv describing the present samples.}

\item{idat.dir}{path where the idat files are located}

\item{sample.col.name}{name of the column in the csv that contains the names of the samples}

\item{preprocess}{whenever rnbeads default preprocessing pipeline should be run. PP results are saved in pp_report/ relatived to workin dir}
}
\value{
rnbeadsSet object
}
\description{
Reads EPIC arrays idat files with RnBeads default pipeline
}
