% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEPICidats.R
\name{readEPIC_RDSgranges}
\alias{readEPIC_RDSgranges}
\title{Reads an Grange object saved in RDS format.}
\usage{
readEPIC_RDSgranges(rdsfile)
}
\arguments{
\item{rdsfile}{to the rnbset object saved on disk}
}
\value{
grange Object
}
\description{
The grange object musst contain (T)otal,(M)ethylated Read counts in the metadata.
}
