% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savePMDSegments.R
\name{savePMDSegments}
\alias{savePMDSegments}
\title{Save PMDs to bed file or RDS}
\usage{
savePMDSegments(PMDs, GRangesFilename = NULL, TableFilename = NULL)
}
\arguments{
\item{PMDs}{GRange segmentation}

\item{GRangesFilename}{path of the GRange output file}

\item{TableFilename}{path of the RDS output file}
}
\description{
Saves PMD segments to disk. Use either GRangesFilename for bed file format or TableFilename for RDS format.
}
