% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentPMDs.R
\name{segmentPMDsKNN}
\alias{segmentPMDsKNN}
\title{Segments a genome into PMD and nonPMD regions using an HMM and KNNs}
\usage{
segmentPMDsKNN(
  m,
  training.chr.sel = "chr2",
  pdfFilename = NULL,
  seqLengths = "hg19",
  knn = c(30, 50, 90),
  q = 1,
  cutoff = c(50000, 1e+05, 150000),
  num.cores = 2,
  markLowDensitySegm = T
)
}
\arguments{
\item{m}{Granges including Metadata for Total,Methylated Counts}

\item{training.chr.sel}{the chromsome that should be used for the training of the HMM}

\item{seqLengths}{a vector representing the sequence length for each chromosome of the assembly}

\item{knn}{a vector of values to be used with KNN}

\item{q}{use quantile based distance cutoff value, for each chromsome, 1=disabled}

\item{cutoff}{a vector of distances to be used for cutoff in KNN}

\item{num.cores}{to be used CPU cores. Multithreading handled by doSNOW cluster}

\item{markLowDensitySegm}{whenever regions with low density points 5\%<,>95\% should be marked seperately}
}
\value{
GRange object with the PMD/nonPMD segments
}
\description{
Estimates PMD segments based on alphaValue estimation combinded with hidden markov model prediction. 
The method is adapted from MethylseekR(MSR) and extended with KNN datapoint selection. 
KNN will choose the neighrest neighorbing points based on the genomic distance. 
These points will be used for alpha value estimation, like in MSR.
Furthermore a distance cutoff is definded such that far away datapoints will not be considered for alpha estimation.
This method is applied for multiple values for K, and all estimated alpha values will be averaged.
The hidden markov model will than be trained and used with those alpha values.
}
