% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeSegmentation.R
\name{writeGranges2bed}
\alias{writeGranges2bed}
\title{Write a GRange object a bed.gz file}
\usage{
writeGranges2bed(granges, seg.file.name)
}
\arguments{
\item{granges}{representing the segmentation}

\item{file.name}{file path and name. Bed.gz will be concatenated}
}
\description{
Write a GRange object a bed.gz file
}
