% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeSegmentation.R
\name{writeNpmds2bed}
\alias{writeNpmds2bed}
\title{Write nonPMDs from a GRange Segmentation object a bed.gz file}
\usage{
writeNpmds2bed(granges, seg.file.name)
}
\arguments{
\item{granges}{representing the segmentation}

\item{file.name}{file path and name. Bed.gz will be concatenated}
}
\description{
Write nonPMDs from a GRange Segmentation object a bed.gz file
}
