% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeSegmentation.R
\name{writePMDs2bed}
\alias{writePMDs2bed}
\title{Write PMDs from a GRange Segmentation object a bed.gz file}
\usage{
writePMDs2bed(granges, seg.file.name)
}
\arguments{
\item{granges}{representing the segmentation}

\item{file.name}{file path and name. Bed.gz will be concatenated}
}
\description{
Write PMDs from a GRange Segmentation object a bed.gz file
}
