% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeSegmentation.R
\name{writeSegmentation}
\alias{writeSegmentation}
\title{Write a PMD/nPMD Segmentation to a bed.gz file}
\usage{
writeSegmentation(granges, seg.file.name, score.col.name = NULL)
}
\arguments{
\item{granges}{representing the segmentation}

\item{score.col.name}{an optional column that contains scores for each segment}

\item{file.name}{file path and name. Bed.gz will be concatenated}
}
\description{
The grange object musst contain (T)otal,(M)ethylated Read counts in the metadata.
}
